Add: json parser

+++ app/src/main/java/com/vswa/data/models/LocationApp.java
@@ -2,6 +2,6 @@ package com.vswa.data.models;
 
 public class LocationApp {
     public String locationName;
-    public Double longitude;
-    public Double latitude;
+    public float longitude;
+    public float latitude;
 }
